+++ screens/dashboardScreen/index.js
@@ -1,0 +1,14 @@
+import { connect } from 'react-redux';
+import DashboardScreen from './DashboardScreen';
+
+const mapStateToProps = state => {
+  return {
+  }
+}
+
+const mapDispatchToProps = dispatch => {
+  return {
+  }
+}
+
+export default connect(mapStateToProps, mapDispatchToProps)(DashboardScreen);
