+++ screens/loadingScreen/LoadingScreen.js
@@ -1,8 +1,9 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
-import { View, Text, StyleSheet, ActivityIndicator } from 'react-native';
+import { View, ActivityIndicator } from 'react-native';
 import firebase from 'firebase';
-import {updateUser} from '../src/actions/userActions'
+import styles from './css';
+
 class LoadingScreen extends Component {
   componentDidMount() {
     this.checkIfLoggedIn();
@@ -11,14 +12,14 @@ class LoadingScreen extends Component {
   checkIfLoggedIn = () => {
     firebase.auth().onAuthStateChanged(
       function(user) {
-        console.log('AUTH STATE CHANGED CALLED ');
+        // console.log('AUTH STATE CHANGED CALLED ');
         if (user) {
           console.log('Update user >>> ', user);
           this.props.updateUser(user);
-          console.log('navigation to >>> Main');
+          // console.log('navigation to >>> Main');
           this.props.navigation.navigate('Main');
         } else {
-          console.log('navigation to >>> LoginScreen');
+          // console.log('navigation to >>> LoginScreen');
           this.props.navigation.navigate('LoginScreen');
         }
       }.bind(this)
@@ -34,24 +35,4 @@ class LoadingScreen extends Component {
   }
 }
 
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    alignItems: 'center',
-    justifyContent: 'center'
-  }
-});
-
-const mapStateToProps = state => {
-  return {
-  }
-}
-
-const mapDispatchToProps = dispatch => {
-  return {
-      updateUser: (user) => dispatch(updateUser(user)),
-  }
-}
-
-export default connect(mapStateToProps, mapDispatchToProps)(LoadingScreen);
-
+export default LoadingScreen;
