+++ screens/loginScreen/LoginScreen.js
@@ -1,38 +1,8 @@
 import React, { Component } from 'react';
-import { createStore } from 'redux';
-import { Provider } from 'react-redux';
-import { View, Text, StyleSheet, Button } from 'react-native';
+import { View, Button } from 'react-native';
 import * as Google from 'expo-google-app-auth';
 import firebase from 'firebase';
-
-/**
- * Store - holds our state - THERE IS ONLY ONE STATE 
- * Action - State can be modified using actions - SIMPLE OBJECTS 
- * Dispatcher - Action needs to be sent by someone - known as dispatching an action
- * Reducer - receives the action and modifies the state to give us a new state 
- *  - pure functions 
- *  - only mandatory argument is the 'type' 
- * Subscriber - listens for state change to update the ui  
- */
-const initialState = {
-  counter: 0,
-  // user: null,
-}
-const reducer = (state = initialState, action) => {
-  switch (action.type) {
-      case 'INCREASE_COUNTER':
-          return { counter: state.counter + 1 }
-      case 'DECREASE_COUNTER':
-          return { counter: state.counter - 1 }
-      // case 'SET_USER':
-      //       console.log('action >>>',action);
-      //       return { user: action.user  }
-      default:
-        return state;
-  }
-  return state
-}
-const store = createStore(reducer)
+import styles from './css';
 
 class LoginScreen extends Component {
 
@@ -147,13 +117,5 @@ class LoginScreen extends Component {
     );
   }
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    alignItems: 'center',
-    justifyContent: 'center'
-  }
-});
 
 export default LoginScreen;
