+++ screens/loginScreen/css.js
@@ -5,9 +5,9 @@ import {
   const styles = StyleSheet.create({
     container: {
       flex: 1,
-      paddingTop: 15,
-      backgroundColor: '#fff',
-    },
+      alignItems: 'center',
+      justifyContent: 'center'
+    }
   });
 
   export default styles;
