+++ screens/loginScreen/index.js
@@ -1,0 +1,14 @@
+import { connect } from 'react-redux';
+import LoginScreen from './LoginScreen';
+
+const mapStateToProps = state => {
+  return {
+  }
+}
+
+const mapDispatchToProps = dispatch => {
+  return {
+  }
+}
+
+export default connect(mapStateToProps, mapDispatchToProps)(LoginScreen);
